Refactor BackButton
