+++ web/components/General/BackButton.tsx
@@ -1,4 +1,5 @@
 import { useRouter } from "next/navigation";
+import { ArrowLeftIcon } from "../Icons";
 
 interface BackButtonProps {
   className?: string;
@@ -6,6 +7,7 @@ interface BackButtonProps {
   goTo?: string;
 }
 
+/** A back button that navigates to the page the user came from or a specified URL. */
 export default function BackButton({
   className = "",
   text = "Back",
@@ -19,24 +21,13 @@ export default function BackButton({
       router.back();
     }
   };
+
   return (
     <button
       onClick={handleClick}
       className={`flex items-center gap-2 text-zinc-600 hover:text-stone-900 transition-colors ${className}`}
     >
-      <svg
-        width="20"
-        height="20"
-        viewBox="0 0 24 24"
-        fill="none"
-        stroke="currentColor"
-        strokeWidth="2"
-        strokeLinecap="round"
-        strokeLinejoin="round"
-      >
-        <path d="m12 19-7-7 7-7" />
-        <path d="M19 12H5" />
-      </svg>
+      <ArrowLeftIcon size={20} />
       <span className="text-sm font-medium">{text}</span>
     </button>
   );
