+++ web/components/PostForm.tsx
@@ -3,7 +3,7 @@
 import { getSupportedImageTypes, validateImage } from "@/utils/imageUtils";
 import Image from "next/image";
 import { useEffect, useRef, useState } from "react";
-import BackButton from "./BackButton";
+import BackButton from "./General/BackButton";
 import ErrorMessage from "./General/ErrorMessage";
 import { ImageIcon } from "./Icons";
